requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
opencv-python-headless>=4.8.0.74
numpy>=1.25.0
torch>=2.0.1
torchvision>=0.15.2
scipy>=1.10.1
pillow>=9.5.0
ultralytics>=8.0.20